:truck: rename grpc package to end in v1_6
