+++ instrumentation/grpc-1.5/src/test/java/io/opentelemetry/javaagent/instrumentation/hypertrace/grpc/v1_6/NoopGreeterService.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.opentelemetry.javaagent.instrumentation.hypertrace.grpc.v1_5;
+package io.opentelemetry.javaagent.instrumentation.hypertrace.grpc.v1_6;
 
 import io.grpc.stub.StreamObserver;
 import org.hypertrace.example.GreeterGrpc;
